🔨 Refactor nikto to separate json and xml (#9680)
* basic structure

* json part

* xml part

* flake8
